+++ working/Makefile
@@ -1,11 +1,11 @@
 CC = gcc
-CFLAGS = -Wall -g -std=c99 -pedantic -Werror
-LDFLAGS = -lm -lSDL2
+CFLAGS = -Wall -g -std=c99 -pedantic -Werror 
+LDFLAGS = -lm
 
 SRC_FILES = $(wildcard *.c)
 OBJ_FILES = $(SRC_FILES:.c=.o)
 
-PROGRAM = UwU
+PROGRAM = uwu
 
 %.o: %.c
 	$(CC) $(CFLAGS) -c $< -o $@ $(LDFLAGS)
@@ -15,3 +15,9 @@ $(PROGRAM): $(OBJ_FILES)
 
 clean:
 	rm -f $(PROGRAM) $(OBJ_FILES)
+
+valgrind:
+	valgrind --leak-check=full --track-origins=yes --show-reachable=yes ./uwu
+
+valgrindsdl:
+	valgrind --suppressions=suppressions_20221_tp1.supp --leak-check=full ./uwu
